fixed eclipse warning
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@830094 13f79535-47bb-0310-9956-ffa450edef68

+++ tutorials/src/org/apache/pivot/tutorials/menus/MenuButtons.java
@@ -137,8 +137,8 @@ public class MenuButtons implements Application {
     public Point getRandomLocation(Shape shape) {
         Bounds bounds = shape.getBounds();
 
-        int x = (int)(Math.random() * (double)(MAX_X - bounds.width));
-        int y = (int)(Math.random() * (double)(MAX_Y - bounds.height));
+        int x = (int)(Math.random() * (MAX_X - bounds.width));
+        int y = (int)(Math.random() * (MAX_Y - bounds.height));
 
         return new Point(x, y);
     }
